requirements.txt
fasttext==0.9.2
gensim==4.1.2
matplotlib==3.4.2
pandas==1.2.5
namu-wiki-extractor==0.2.2
tweepy==3.10.0
soynlp==0.0.493
konlpy==0.5.2
kss==2.5.1
tqdm==4.62.3
# tensorflow==2.3.0    # GTX 10, RTX 20 시리즈 GPU 및 CUDA 10.1 환경에서 작동 확인함. 자신의 환경에 맞는 버전 설치할 것.
